Move studies caching to Transmart resource service.
