+++ src/app/services/resource.service.spec.ts
@@ -3,8 +3,6 @@ import {TestBed, inject} from '@angular/core/testing';
 import {ResourceService} from './resource.service';
 import {TransmartResourceService} from './transmart-services/transmart-resource.service';
 import {TransmartResourceServiceMock} from './mocks/transmart-resource.service.mock';
-import {TransmartStudiesService} from './transmart-services/transmart-studies.service';
-import {TransmartStudiesServiceMock} from './mocks/transmart-studies.service.mock';
 
 describe('ResourceService', () => {
   let resourceService: ResourceService;
@@ -17,10 +15,6 @@ describe('ResourceService', () => {
         {
           provide: TransmartResourceService,
           useClass: TransmartResourceServiceMock
-        },
-        {
-          provide: TransmartStudiesService,
-          useClass: TransmartStudiesServiceMock
         }
       ]
     });
